Introduce Proper OOP Principles
