--- BasicPoseModule.py
@@ -1,68 +1,0 @@
-import cv2
-import mediapipe as mp
-
-class poseDetector() :
-    
-    def __init__(self, mode=False, complexity=1, smooth_landmarks=True,
-                 enable_segmentation=False, smooth_segmentation=True,
-                 detectionCon=0.5, trackCon=0.5):
-        
-        self.mode = mode 
-        self.complexity = complexity
-        self.smooth_landmarks = smooth_landmarks
-        self.enable_segmentation = enable_segmentation
-        self.smooth_segmentation = smooth_segmentation
-        self.detectionCon = detectionCon
-        self.trackCon = trackCon
-        
-        self.mpDraw = mp.solutions.drawing_utils
-        self.mpPose = mp.solutions.pose
-        self.pose = self.mpPose.Pose(self.mode, self.complexity, self.smooth_landmarks,
-                                     self.enable_segmentation, self.smooth_segmentation,
-                                     self.detectionCon, self.trackCon)
-        
-        
-    def findPose (self, img, draw=True):
-        imgRGB = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-        self.results = self.pose.process(imgRGB)
-        
-        if self.results.pose_landmarks:
-            if draw:
-                self.mpDraw.draw_landmarks(img,self.results.pose_landmarks,
-                                           self.mpPose.POSE_CONNECTIONS)
-                
-        return img
-    
-    def findPosition(self, img, draw=True):
-        lmList = []
-        if self.results.pose_landmarks:
-            for id, lm in enumerate(self.results.pose_landmarks.landmark):
-                #finding height, width of the image printed
-                h, w, c = img.shape
-                #Determining the pixels of the landmarks
-                cx, cy = int(lm.x * w), int(lm.y * h)
-                lmList.append([id, cx, cy])
-                if draw:
-                    cv2.circle(img, (cx, cy), 5, (255,0,0), cv2.FILLED)
-        return lmList
-    
-
-        
-        
-        
-def main():
-    detector = poseDetector()
-    cap = cv2.VideoCapture(0)
-    while cap.isOpened():
-        ret, img = cap.read() #ret is just the return variable, not much in there that we will use. 
-        if ret:    
-            img = detector.findPose(img)
-            cv2.imshow('Pose Detection', img)
-        if cv2.waitKey(10) & 0xFF == ord('q'):
-            break
-            
-    cap.release()
-    cv2.destroyAllWindows()
-    
-if __name__ == "__main__":
-    main()
