+++ base.py
@@ -1,14 +1,11 @@
 import cv2
 import mediapipe as mp
-import math
 
-class poseDetector() :
-    
+class BasePoseDetector:
     def __init__(self, mode=False, complexity=1, smooth_landmarks=True,
                  enable_segmentation=False, smooth_segmentation=True,
                  detectionCon=0.5, trackCon=0.5):
-        
-        self.mode = mode 
+        self.mode = mode
         self.complexity = complexity
         self.smooth_landmarks = smooth_landmarks
         self.enable_segmentation = enable_segmentation
@@ -23,14 +20,13 @@ class poseDetector() :
                                      self.detectionCon, self.trackCon)
         
         
-    def findPose (self, img, draw=True):
-        imgRGB = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+    def findPose(self, img, draw=True):
+        imgRGB = img[:, :, ::-1]
         self.results = self.pose.process(imgRGB)
         
-        if self.results.pose_landmarks:
-            if draw:
-                self.mpDraw.draw_landmarks(img,self.results.pose_landmarks,
-                                           self.mpPose.POSE_CONNECTIONS)
+        if self.results.pose_landmarks and draw:
+            self.mpDraw.draw_landmarks(img,self.results.pose_landmarks,
+                    self.mpPose.POSE_CONNECTIONS)
                 
         return img
     
@@ -46,47 +42,14 @@ class poseDetector() :
                 if draw:
                     cv2.circle(img, (cx, cy), 5, (255,0,0), cv2.FILLED)
         return self.lmList
-        
-    def findAngle(self, img, p1, p2, p3, draw=True):   
-        #Get the landmarks
-        x1, y1 = self.lmList[p1][1:]
-        x2, y2 = self.lmList[p2][1:]
-        x3, y3 = self.lmList[p3][1:]
-        
-        #Calculate Angle
-        angle = math.degrees(math.atan2(y3-y2, x3-x2) - 
-                             math.atan2(y1-y2, x1-x2))
-        if angle < 0:
-            angle += 360
-            if angle > 180:
-                angle = 360 - angle
-        elif angle > 180:
-            angle = 360 - angle
-        # print(angle)
-        
-        #Draw
-        if draw:
-            cv2.line(img, (x1, y1), (x2, y2), (255,255,255), 3)
-            cv2.line(img, (x3, y3), (x2, y2), (255,255,255), 3)
 
-            
-            cv2.circle(img, (x1, y1), 5, (0,0,255), cv2.FILLED)
-            cv2.circle(img, (x1, y1), 15, (0,0,255), 2)
-            cv2.circle(img, (x2, y2), 5, (0,0,255), cv2.FILLED)
-            cv2.circle(img, (x2, y2), 15, (0,0,255), 2)
-            cv2.circle(img, (x3, y3), 5, (0,0,255), cv2.FILLED)
-            cv2.circle(img, (x3, y3), 15, (0,0,255), 2)
-            
-            cv2.putText(img, str(int(angle)), (x2-50, y2+50), 
-                        cv2.FONT_HERSHEY_PLAIN, 2, (0,0,255), 2)
-        return angle
-        
 
 def main():
-    detector = poseDetector()
+    detector = BasePoseDetector()
     cap = cv2.VideoCapture(0)
     while cap.isOpened():
-        ret, img = cap.read() #ret is just the return variable, not much in there that we will use. 
+        ret, img = cap.read()
+        # ret is just the return variable, not much in there that we will use. 
         if ret:    
             img = detector.findPose(img)
             cv2.imshow('Pose Detection', img)
