+++ pose.py
@@ -1,0 +1,57 @@
+import cv2
+import math
+from base import BasePoseDetector
+
+class PoseDetector(BasePoseDetector):
+        
+    def findAngle(self, img, p1, p2, p3, draw=True):   
+        #Get the landmarks
+        x1, y1 = self.lmList[p1][1:]
+        x2, y2 = self.lmList[p2][1:]
+        x3, y3 = self.lmList[p3][1:]
+        
+        #Calculate Angle
+        angle = math.degrees(math.atan2(y3-y2, x3-x2) - 
+                             math.atan2(y1-y2, x1-x2))
+        if angle < 0:
+            angle += 360
+            if angle > 180:
+                angle = 360 - angle
+        elif angle > 180:
+            angle = 360 - angle
+        # print(angle)
+        
+        #Draw
+        if draw:
+            cv2.line(img, (x1, y1), (x2, y2), (255,255,255), 3)
+            cv2.line(img, (x3, y3), (x2, y2), (255,255,255), 3)
+
+            
+            cv2.circle(img, (x1, y1), 5, (0,0,255), cv2.FILLED)
+            cv2.circle(img, (x1, y1), 15, (0,0,255), 2)
+            cv2.circle(img, (x2, y2), 5, (0,0,255), cv2.FILLED)
+            cv2.circle(img, (x2, y2), 15, (0,0,255), 2)
+            cv2.circle(img, (x3, y3), 5, (0,0,255), cv2.FILLED)
+            cv2.circle(img, (x3, y3), 15, (0,0,255), 2)
+            
+            cv2.putText(img, str(int(angle)), (x2-50, y2+50), 
+                        cv2.FONT_HERSHEY_PLAIN, 2, (0,0,255), 2)
+        return angle
+        
+
+def main():
+    detector = PoseDetector()
+    cap = cv2.VideoCapture(0)
+    while cap.isOpened():
+        ret, img = cap.read() 
+        if ret:    
+            img = detector.findPose(img)
+            cv2.imshow('Pose Detection', img)
+        if cv2.waitKey(10) & 0xFF == ord('q'):
+            break
+            
+    cap.release()
+    cv2.destroyAllWindows()
+    
+if __name__ == "__main__":
+    main()
