+++ requirements.txt
@@ -1,0 +1,3 @@
+numpy
+mediapipe
+opencv-python
